clean Main.cs and Globals.cs

+++ BravoGame/Source/Engine/Globals.cs
@@ -34,7 +34,7 @@ namespace BravoGame
             if (Pos.Y - focus.Y != 0)
             {
                 h = (float)Math.Sqrt(Math.Pow(Pos.X - focus.X, 2) + Math.Pow(Pos.Y - focus.Y, 2));
-                sineTheta = (float)(Math.Abs(Pos.Y - focus.Y) / h); // * ((item.Pos.Y-focus.Y)/(Math.Abs(item.Pos.Y-focus.Y))));
+                sineTheta = Math.Abs(Pos.Y - focus.Y) / h; // * ((item.Pos.Y-focus.Y)/(Math.Abs(item.Pos.Y-focus.Y))));
             }
             else
             {
